Add Rank Tracking metric filters and fix null-last sort ordering
