+++ src/client/features/rank-tracking/RankTrackingFilters.tsx
@@ -1,56 +1,12 @@
 import { RotateCcw } from "lucide-react";
-import { LOCATIONS } from "@/client/features/keywords/locations";
-import { devicesLabel } from "@/shared/rank-tracking";
-import type {
-  RankTrackingConfig,
-  RankTrackingRow,
-} from "@/types/schemas/rank-tracking";
+import type { DomainListFilters, Filters } from "./RankTrackingFilters.logic";
 
-export type Filters = {
-  include: string;
-  exclude: string;
-  minDesktopPos: string;
-  maxDesktopPos: string;
-  minMobilePos: string;
-  maxMobilePos: string;
-};
-
-type DomainFilterableConfig = Pick<
-  RankTrackingConfig,
-  "domain" | "devices" | "locationCode"
->;
-
-export type DomainListFilters = {
-  query: string;
-  device: "all" | RankTrackingConfig["devices"];
-  locationCode: string;
-};
+export * from "./RankTrackingFilters.logic";
 
 type DomainListFilterOption = {
   value: string;
   label: string;
 };
-
-export const EMPTY_FILTERS: Filters = {
-  include: "",
-  exclude: "",
-  minDesktopPos: "",
-  maxDesktopPos: "",
-  minMobilePos: "",
-  maxMobilePos: "",
-};
-
-export const EMPTY_DOMAIN_LIST_FILTERS: DomainListFilters = {
-  query: "",
-  device: "all",
-  locationCode: "all",
-};
-
-const DEVICE_FILTER_ORDER: RankTrackingConfig["devices"][] = [
-  "both",
-  "desktop",
-  "mobile",
-];
 
 export function FilterPanel({
   filters,
@@ -124,6 +80,29 @@ export function FilterPanel({
           maxValue={filters.maxMobilePos}
           onMinChange={(v) => update("minMobilePos", v)}
           onMaxChange={(v) => update("maxMobilePos", v)}
+        />
+      </div>
+      <div className="grid grid-cols-1 gap-3 lg:grid-cols-3">
+        <RangeFilter
+          title="Volume"
+          minValue={filters.minVolume}
+          maxValue={filters.maxVolume}
+          onMinChange={(v) => update("minVolume", v)}
+          onMaxChange={(v) => update("maxVolume", v)}
+        />
+        <RangeFilter
+          title="Keyword difficulty"
+          minValue={filters.minKd}
+          maxValue={filters.maxKd}
+          onMinChange={(v) => update("minKd", v)}
+          onMaxChange={(v) => update("maxKd", v)}
+        />
+        <RangeFilter
+          title="CPC"
+          minValue={filters.minCpc}
+          maxValue={filters.maxCpc}
+          onMinChange={(v) => update("minCpc", v)}
+          onMaxChange={(v) => update("maxCpc", v)}
         />
       </div>
     </div>
@@ -261,139 +240,4 @@ function RangeFilter({
       </div>
     </div>
   );
-}
-
-export function applyDomainListFilters<T extends DomainFilterableConfig>(
-  configs: T[],
-  filters: DomainListFilters,
-): T[] {
-  const query = filters.query.trim().toLowerCase();
-  const locationCode =
-    filters.locationCode === "all" ? null : Number(filters.locationCode);
-
-  return configs.filter((config) => {
-    if (query && !config.domain.toLowerCase().includes(query)) return false;
-
-    if (filters.device !== "all" && config.devices !== filters.device) {
-      return false;
-    }
-
-    if (locationCode !== null && config.locationCode !== locationCode) {
-      return false;
-    }
-
-    return true;
-  });
-}
-
-export function getDomainListFilterOptions(configs: DomainFilterableConfig[]): {
-  devices: DomainListFilterOption[];
-  locations: DomainListFilterOption[];
-} {
-  const deviceValues = new Set(configs.map((config) => config.devices));
-  const devices = DEVICE_FILTER_ORDER.filter((device) =>
-    deviceValues.has(device),
-  ).map((device) => ({
-    value: device,
-    label: devicesLabel(device),
-  }));
-
-  const locationMap = new Map<number, string>();
-  for (const config of configs) {
-    locationMap.set(
-      config.locationCode,
-      LOCATIONS[config.locationCode] ?? String(config.locationCode),
-    );
-  }
-
-  const locations = Array.from(locationMap, ([code, label]) => ({
-    value: String(code),
-    label,
-  })).toSorted((a, b) => a.label.localeCompare(b.label));
-
-  return { devices, locations };
-}
-
-export function applyFilters(
-  rows: RankTrackingRow[],
-  filters: Filters,
-): RankTrackingRow[] {
-  const includeTerms = filters.include
-    ? filters.include
-        .toLowerCase()
-        .split(",")
-        .map((t) => t.trim())
-        .filter(Boolean)
-    : [];
-  const excludeTerms = filters.exclude
-    ? filters.exclude
-        .toLowerCase()
-        .split(",")
-        .map((t) => t.trim())
-        .filter(Boolean)
-    : [];
-
-  return rows.filter((row) => {
-    const kw = row.keyword.toLowerCase();
-
-    if (includeTerms.length > 0 && !includeTerms.some((t) => kw.includes(t)))
-      return false;
-
-    if (excludeTerms.some((t) => kw.includes(t))) return false;
-
-    if (
-      !matchesPositionFilter(
-        row.desktop.position,
-        filters.minDesktopPos,
-        filters.maxDesktopPos,
-      )
-    )
-      return false;
-
-    if (
-      !matchesPositionFilter(
-        row.mobile.position,
-        filters.minMobilePos,
-        filters.maxMobilePos,
-      )
-    )
-      return false;
-
-    return true;
-  });
-}
-
-export function matchesPositionFilter(
-  position: number | null,
-  minValue: string,
-  maxValue: string,
-): boolean {
-  if (!minValue && !maxValue) return true;
-
-  const max = maxValue === "" ? Infinity : Number(maxValue);
-  if (max === 0) return position === null;
-
-  if (position === null) return false;
-
-  const min = minValue === "" ? 0 : Number(minValue);
-  return position >= min && position <= max;
-}
-
-export function countActiveFilters(filters: Filters): number {
-  let count = 0;
-  if (filters.include) count++;
-  if (filters.exclude) count++;
-  if (filters.minDesktopPos || filters.maxDesktopPos) count++;
-  if (filters.minMobilePos || filters.maxMobilePos) count++;
-  return count;
-}
-
-export function countActiveDomainListFilters(
-  filters: DomainListFilters,
-): number {
-  let count = 0;
-  if (filters.query.trim()) count++;
-  if (filters.device !== "all") count++;
-  if (filters.locationCode !== "all") count++;
-  return count;
 }
